Consolidate unwrap (#2523)
